refactor test folder to match view folders
