+++ tests/test_views/admin/settings/catalog/operations/test_create.py
