+++ tests/test_views/admin/settings/catalog/operations/test_edit.py
