+++ tests/test_views/admin/settings/catalog/services/test_edit.py
