+++ tests/test_views/admin/settings/tower/base_test_tower.py
