+++ tests/test_views/admin/settings/tower/test_delete.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 
 from service_catalog.models import TowerServer, JobTemplate
-from tests.test_views.admin.tower.base_test_tower import BaseTestTower
+from tests.test_views.admin.settings.tower.base_test_tower import BaseTestTower
 
 
 class AdminTowerDeleteViewsTest(BaseTestTower):
