+++ tests/test_views/admin/settings/tower/test_get_details.py
@@ -3,7 +3,7 @@ from unittest import mock
 
 from django.urls import reverse
 
-from tests.test_views.admin.tower.base_test_tower import BaseTestTower
+from tests.test_views.admin.settings.tower.base_test_tower import BaseTestTower
 
 
 class AdminTowerGetViewsTest(BaseTestTower):
