+++ tests/test_views/admin/settings/tower/test_list.py
@@ -1,6 +1,6 @@
 from django.urls import reverse
 
-from tests.test_views.admin.tower.base_test_tower import BaseTestTower
+from tests.test_views.admin.settings.tower.base_test_tower import BaseTestTower
 
 
 class AdminTowerListViewsTest(BaseTestTower):
